deployment/requirements.txt
streamlit
pandas
pillow
numpy
tensorflow<2.16
